+++ frontend/app/mocks/node.ts
@@ -1,6 +1,5 @@
 import { setupServer } from 'msw/node';
 
-import { getClientApplicationApiMockHandlers } from './client-application-api.server';
 import { getCCTApiMockHandlers } from '~/mocks/cct-api.server';
 import { getPowerPlatformApiMockHandlers } from '~/mocks/power-platform-api.server';
 import { getRaoidcMockHandlers } from '~/mocks/raoidc.server';
@@ -14,5 +13,4 @@ export const server = setupServer(
   ...(mockEnabled('raoidc') ? getRaoidcMockHandlers() : []),
   ...(mockEnabled('status-check') ? getStatusCheckApiMockHandlers() : []),
   ...(mockEnabled('wsaddress') ? getWSAddressApiMockHandlers() : []),
-  ...(mockEnabled('client-application') ? getClientApplicationApiMockHandlers() : []),
 );
